requirements.txt
# Vision dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.41.0         # Florence-2 support
einops>=0.7.0                # required by Florence-2
timm>=0.9.12                 # required by Florence-2
Pillow>=10.0.0
requests>=2.28.0

# Vision server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# MuJoCo (optional – only needed for live scene capture)
# mujoco>=3.1.0

# Dev / testing
# pytest>=8.0
